add file existance checks for csharp setupper

+++ fish/setup/csharp.fish
@@ -2,8 +2,16 @@
 
 set -l choreTrackerBrokenSymlink /mnt/c/Programming/binaries/ChoreTracker.exe
 set -l choreTrackerTarget /mnt/c/Programming/csproj/ChoreTracker/ChoreTracker/bin/Debug/net7.0/ChoreTracker.exe
-ln -sf $choreTrackerTarget $choreTrackerBrokenSymlink
+if test -e $choreTrackerTarget
+	ln -sf $choreTrackerTarget $choreTrackerBrokenSymlink
+else
+	echo 'the target "'$choreTrackerTarget '" doesn\'t exist!'
+end
 
 set -l weldeBrokenSymlink /mnt/c/Programming/binaries/Welde.exe
 set -l weldeTarget /mnt/c/Programming/csproj/Welde/Welde/bin/Debug/net7.0/Welde.exe
-ln -sf $weldeTarget $weldeBrokenSymlink
+if test -e $weldeTarget
+	ln -sf $weldeTarget $weldeBrokenSymlink
+else
+	echo 'the target "'$weldeTarget'" doesn\'t exist!'
+end
